requirements.txt
pytest
pytest_bdd
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
lxml-4.3.1-cp34-cp34m-win_amd64
pytest-allure-adaptor